Minor code refactoring, created an interface for Collision detection
