+++ src/main/java/org/example/CollisionDetection.java
@@ -1,0 +1,6 @@
+package org.example;
+
+public interface CollisionDetection {
+  public void checkEdges(int screenWidth, int screenHeight);
+  public boolean collidesWith(Sprite other);
+}
